fix(logger): unify LoggerPort protocol to eliminate duplicates
- Moved LoggerPort definition to a new unified location in `domain/protocols/logger.py` to serve as a single source of truth.
- Updated all imports across the codebase to reference the new LoggerPort location, ensuring consistency and preventing TypeErrors.
- Removed outdated LoggerPort definitions from `domain/ports/base.py` and `domain/abnormal/ports.py` to streamline the code structure.
- Enhanced LoggerPort with a flexible signature to support various logging contexts.
